limit ffmpeg sample video creation threads and other minor changes

+++ bot/helper/mirror_utils/status_utils/extract_status.py
@@ -1,6 +1,6 @@
 from time import time
 
-from bot import LOGGER
+from bot import LOGGER, subprocess_lock
 from bot.helper.ext_utils.status_utils import (
     get_readable_file_size,
     MirrorStatus,
@@ -65,8 +65,12 @@ class ExtractStatus:
 
     async def cancel_task(self):
         LOGGER.info(f"Cancelling Extract: {self.listener.name}")
-        if self.listener.suproc is not None:
-            self.listener.suproc.kill()
-        else:
-            self.listener.suproc = "cancelled"
+        async with subprocess_lock:
+            if (
+                self.listener.suproc is not None
+                and self.listener.suproc.returncode is None
+            ):
+                self.listener.suproc.kill()
+            else:
+                self.listener.suproc = "cancelled"
         await self.listener.onUploadError("extracting stopped by user!")
