+++ bot/helper/mirror_utils/status_utils/split_status.py
@@ -1,4 +1,4 @@
-from bot import LOGGER
+from bot import LOGGER, subprocess_lock
 from bot.helper.ext_utils.status_utils import get_readable_file_size, MirrorStatus
 
 
@@ -37,8 +37,12 @@ class SplitStatus:
 
     async def cancel_task(self):
         LOGGER.info(f"Cancelling Split: {self.listener.name}")
-        if self.listener.suproc is not None:
-            self.listener.suproc.kill()
-        else:
-            self.listener.suproc = "cancelled"
+        async with subprocess_lock:
+            if (
+                self.listener.suproc is not None
+                and self.listener.suproc.returncode is None
+            ):
+                self.listener.suproc.kill()
+            else:
+                self.listener.suproc = "cancelled"
         await self.listener.onUploadError("splitting stopped by user!")
